# Optimize scalar inversion by implementing square()