This shows an 11% speedup for invert(). From `cargo bench`:

* Old `invert`: `61,945 ns/iter (+/- 19,909)`
* New `invert`: `55,051 ns/iter (+/- 6,177)`
* scalar_unpacked_multiply_add: `181 ns/iter (+/- 7)`
* square: `159 ns/iter (+/- 21)`

This addresses bullet point `b` of https://github.com/isislovecruft/curve25519-dalek/issues/68 (and possibly `c` too?)

Questions I don't know the answer to: should `square` be part of the public API? Should I add a wrapper to `Scalar` instead of only having it on `UnpackedScalar`?

This is my first PR to this repo, I would welcome advice on how to improve 😊