@hdevalence 2017-08-25
This looks good to me.

Re: the API question, I'm not super opinionated, but I think it makes sense for people who care about scalar arithmetic performance (the rare case) to just use the UnpackedScalar type directly.