@hdevalence 2017-08-25
Merged into the `develop` branch instead of `master`.